# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Octo Protocol API
  version: 1.0.0
  description: OpenAPI contract for the /v1 Octo Protocol routes.
servers:
  - url: /
paths:
...
  /v1/wallets/{id}/api-key:
    post:
      summary: Generate or rotate the wallet's API key
      description: >
        Dashboard JWT only. The plaintext key is returned once; only its SHA-256 hash is stored.
        The first key needs no body. Once a key exists, `confirm: true` is required to rotate it
        (which immediately invalidates the previous key); without it the call returns 409.
      operationId: generateApiKey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                confirm:
                  type: boolean
                  description: Must be `true` to rotate an existing key.
      responses:
        "201":
          description: Created — the full key, shown once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        "409":
          description: A key already exists and `confirm` was not `true`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
...
components:
  schemas:
    Envelope:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
...